Fixed Issues @64. 'appId' not found

+++ web/view/scan.py
@@ -170,7 +170,8 @@ def scan():
     if device == 'ios':
         pii_fpath = sc.dump_path(ser, 'Device_Info')
         print('Revelant info saved to db. Deleting {} now.'.format(pii_fpath))
-        cmd = os.unlink(pii_fpath)
+        if os.path.exists(pii_fpath):
+            cmd = os.unlink(pii_fpath)
         # s = catch_err(run_command(cmd), msg="Delete pii failed", cmd=cmd)
         print('iOS PII deleted.')
